feat:create UserInfo component and fetch info from API
